Move mock_environment_dict into own file in recordprocessor
