+++ recordprocessor/tests/test_convert_to_fhir_imms_decorators.py
@@ -16,10 +16,8 @@ from tests.utils_for_recordprocessor_tests.decorator_constants import (
     ExtensionItems,
     RSV_TARGET_DISEASE_ELEMENT,
 )
-from tests.utils_for_recordprocessor_tests.values_for_recordprocessor_tests import (
-    MockFieldDictionaries,
-    MOCK_ENVIRONMENT_DICT,
-)
+from tests.utils_for_recordprocessor_tests.values_for_recordprocessor_tests import MockFieldDictionaries
+from tests.utils_for_recordprocessor_tests.mock_environment_variables import MOCK_ENVIRONMENT_DICT
 
 with patch("os.environ", MOCK_ENVIRONMENT_DICT):
     from constants import Urls
