+++ recordprocessor/tests/test_convert_to_fhir_imms_resource.py
@@ -6,8 +6,8 @@ from unittest.mock import patch
 from tests.utils_for_recordprocessor_tests.values_for_recordprocessor_tests import (
     MockFhirImmsResources,
     MockFieldDictionaries,
-    MOCK_ENVIRONMENT_DICT,
 )
+from tests.utils_for_recordprocessor_tests.mock_environment_variables import MOCK_ENVIRONMENT_DICT
 
 with patch("os.environ", MOCK_ENVIRONMENT_DICT):
     # Do not attempt 'from src.mappings import Vaccine' as this imports a different instance of Vaccine
