+++ recordprocessor/tests/test_make_and_upload_ack_file.py
@@ -2,7 +2,7 @@
 
 import unittest
 from unittest.mock import patch
-from tests.utils_for_recordprocessor_tests.values_for_recordprocessor_tests import MOCK_ENVIRONMENT_DICT
+from tests.utils_for_recordprocessor_tests.mock_environment_variables import MOCK_ENVIRONMENT_DICT
 
 with patch("os.environ", MOCK_ENVIRONMENT_DICT):
     from make_and_upload_ack_file import make_ack_data
