+++ recordprocessor/tests/test_map_target_disease.py
@@ -3,10 +3,8 @@
 import unittest
 from unittest.mock import patch
 from typing import List
-from tests.utils_for_recordprocessor_tests.values_for_recordprocessor_tests import (
-    TargetDiseaseElements,
-    MOCK_ENVIRONMENT_DICT,
-)
+from tests.utils_for_recordprocessor_tests.values_for_recordprocessor_tests import TargetDiseaseElements
+from tests.utils_for_recordprocessor_tests.mock_environment_variables import MOCK_ENVIRONMENT_DICT
 
 with patch("os.environ", MOCK_ENVIRONMENT_DICT):
     from mappings import map_target_disease, Vaccine
