+++ recordprocessor/tests/utils_for_recordprocessor_tests/values_for_recordprocessor_tests.py
@@ -1,28 +1,15 @@
 """Values for use in tests"""
 
+from unittest.mock import patch
 import json
 from decimal import Decimal
-from enum import Enum
+from tests.utils_for_recordprocessor_tests.mock_environment_variables import MOCK_ENVIRONMENT_DICT
+
+with patch("os.environ", MOCK_ENVIRONMENT_DICT):
+    from constants import Urls
+    from mappings import Vaccine
 
 REGION_NAME = "eu-west-2"
-
-
-class Urls:
-    """Urls"""
-
-    SNOMED = "http://snomed.info/sct"
-    NHS_NUMBER = "https://fhir.nhs.uk/Id/nhs-number"
-    NULL_FLAVOUR_CODES = "http://terminology.hl7.org/CodeSystem/v3-NullFlavor"
-    VACCINATION_PROCEDURE = "https://fhir.hl7.org.uk/StructureDefinition/Extension-UKCore-VaccinationProcedure"
-
-
-class Vaccine(Enum):
-    """Disease Codes"""
-
-    COVID_19: str = "COVID19"
-    FLU: str = "FLU"
-    MMR: str = "MMR"
-    RSV: str = "RSV"
 
 
 class MockUniqueIdUris:
@@ -50,38 +37,6 @@ class MockLocalIds:
     COVID19_001_RAVS = create_local_id(MockUniqueIds.COVID19_001, MockUniqueIdUris.RAVS)
     RSV_001_RAVS = create_local_id(MockUniqueIds.RSV_001, MockUniqueIdUris.RAVS)
     RSV_002_RAVS = create_local_id(MockUniqueIds.RSV_002, MockUniqueIdUris.RAVS)
-
-
-class BucketNames:
-    """Class containing bucket names for use in tests"""
-
-    SOURCE = "immunisation-batch-internal-dev-data-sources"
-    DESTINATION = "immunisation-batch-internal-dev-data-destinations"
-    MOCK_FIREHOSE = "mock-firehose-bucket"
-
-
-class Kinesis:
-    """Class containing Kinesis values for use in tests"""
-
-    STREAM_NAME = "imms-batch-internal-dev-processingdata-stream"
-
-
-class Firehose:
-    """Class containing Firehose values for use in tests"""
-
-    STREAM_NAME = "immunisation-fhir-api-internal-dev-splunk-firehose"
-
-
-MOCK_ENVIRONMENT_DICT = {
-    "ENVIRONMENT": "internal-dev",
-    "LOCAL_ACCOUNT_ID": "123456789012",
-    "SOURCE_BUCKET_NAME": BucketNames.SOURCE,
-    "ACK_BUCKET_NAME": BucketNames.DESTINATION,
-    "SHORT_QUEUE_PREFIX": "imms-batch-internal-dev",
-    "KINESIS_STREAM_NAME": Kinesis.STREAM_NAME,
-    "KINESIS_STREAM_ARN": f"arn:aws:kinesis:{REGION_NAME}:123456789012:stream/{Kinesis.STREAM_NAME}",
-    "FIREHOSE_STREAM_NAME": Firehose.STREAM_NAME,
-}
 
 
 class TargetDiseaseElements:
